+++ Drs.Model/Client/ClientPhoneModel.cs
@@ -2,9 +2,9 @@
 {
     public class ClientPhoneModel
     {
-        public int ClientPhoneId { get; set; }
+        public long ClientPhoneId { get; set; }
         public string Phone { get; set; }
-        public int ClientId { get; set; }
+        public long ClientId { get; set; }
 
     }
 }
