+++ Drs.Model/Order/PhoneModel.cs
@@ -4,7 +4,7 @@ namespace Drs.Model.Order
 {
     public class PhoneModel : RecordModel
     {
-        public int PhoneId { get; set; }
+        public long PhoneId { get; set; }
         public string Phone { get; set; }
     }
 }
