+++ Drs.Model/Shared/ListItemModel.cs
@@ -2,7 +2,7 @@
 {
     public class ListItemModel
     {
-        public int? IdKey { get; set; }
+        public long? IdKey { get; set; }
         public string Key { get; set; }
         public string Value { get; set; }
     }
