+++ Drs.ViewModel/Track/OrdersListVm.cs
@@ -171,7 +171,7 @@ namespace Drs.ViewModel.Track
         }
 
 
-        public void OnClientNameChanged(int clientId)
+        public void OnClientNameChanged(long clientId)
         {
             _pagerCache.SearchObjValue = clientId;
             _pagerCache.SearchType = SettingsData.Constants.TrackConst.SEARCH_BY_CLIENTNAME;
@@ -183,13 +183,13 @@ namespace Drs.ViewModel.Track
         {
             OnStatusChanged(SettingsData.Constants.TrackConst.SEARCH_ORDERLIST_ON_PROGRESS, "Buscando pedidos...");
 
-            var pagerDto = new PagerDto<int>
+            var pagerDto = new PagerDto<long>
             {
-                Data = (int)_pagerCache.SearchObjValue,
+                Data = (long)_pagerCache.SearchObjValue,
                 Pager = Pager.PagerModel,
             };
 
-            _client.ExecutionProxy.ExecuteRequest<PagerDto<int>, PagerDto<int>, ResponseMessageData<TrackOrderDto>, ResponseMessageData<TrackOrderDto>>
+            _client.ExecutionProxy.ExecuteRequest<PagerDto<long>, PagerDto<long>, ResponseMessageData<TrackOrderDto>, ResponseMessageData<TrackOrderDto>>
                     (pagerDto, TransferDto.SameType, SharedConstants.Server.TRACK_HUB,
                         SharedConstants.Server.SEARCH_BY_CLIENTNAME_TRACK_HUB_METHOD, TransferDto.SameType)
                         .Subscribe(OnResultSearchOk, OnResultSearchError);
